Making small refactorings once at a time.
In my experience, we should never find ourselves with a non-compilable project for longer than a few minutes while we refactor. This means it must be done very slowly and carefully.

Otherwise, you might be stuck for days with a version of the program that cannot be compiled, and that's trouble - for instance, you cannot run unit tests, you cannot safely pull any version from the veersion control and, of course, you cannot commit your chnages