I do not want the control that shows a dream entry to handle the actual editing/deletion process.
In the original version, it even had a reference to the main form.

Now, how can I unit test the view model when I want to delete or edit if I must have an instance of the main form?

Actually, I want to delegate to the dream diary, that's the class that should handle persistence or anything else.

I created a bus interface that will handle this - how I will implement it, I don't know yet, but I can already mock it and see that the view model only does what falls within its responsibility, and delegates to someone else.

Right now the application will fail during runtime because Ninject cannot bind to it, but we will solve that shortly.

However, I can already run unit tests, and I have broken the dependency that made unit testing impossible