Successfully removed DreamSaveLoadService from the FrmMain form, thus decoupling how we retrieve or persist data from the interface - for now.
Now the problem is that DreamSaveLoadService is using threads and giving cross-threads errors when dealing with the GUI.

I don't want to depend on this, the GUI when receiving events from the viewmodel shouldn't worry about invoking in the main thread.

There is even another question. Is it really worth it at this stage at least to bother with threads instead of blocking the GUI for a few seconds?

AT nay rate, I want to separate the persistence service implementation (in XML in this case) from any multthreading concerns.

Therefore, DreamSaveLoadService will be split: DreamSaveLoadService will be concerned with using a BackgroundWorker to use paralellism while avoiding cross thread problems, while a new PersistenceService will be created that deals with persistence in any way we see fit, without worrying about threads.

Thus, it can be unit tested to some extend - i.e. as long as we can abstract away actually writing to the file system