Finished refactoring SaveLoadService to use a background worker and decouple a technical matter (threads) from the calls we do to the view model from the main view.
My next refactoring target is DreamTagStatistics.

As per resharper it has some unused methods and it has some nasty flaws such as, on the one side, dealing with statistics (an application-side concern) and also with presentation (there are a bunch of presentation fields such as LeftTab, RtfHeader, etc.) We must split both concerns.

Extracting statistics should be come from one class, and how to present them on one or more presenter classes