Small refactoring.
I don't like the Xml writers and readers. It is actually a hand written implementation.

Nowadays I know better than this and there are many ways to find libraries to support this.

I might go as far as saying I can just serialize the dream diary class with everything contained, or I can do something else, but I won't be bothering with technical details and leave the classes as they are for now