More refactoring, this time concerning the settings/last opened locations.
It's funny how I unintentionally introduced a legacy dependency by creating a separate library (EvilTools) that I have no access to.

Therefore, this one is like a legacy system that cannot be tested and thus I have to isolate it completely by creating a wrapper that "interprets" the data coming and going to it so that I have control over it. DreamDiarySettings and IDreamDiarySettings implement this anti-corruption layer pattern