Removed the search interaction logic from FrmMain, so that the interactions are more abstract - FrmMain doesn't know how the search window is managed, and it doesn't even know when to filter out entires or when to show all - this is delegated to the dream diary view model.
Also, I need to pass a "Provider" of entries to the search window because there is a very big possible bug: we pass the dream entries once, and after that they are never updatd no matter how many times we open the window.

Therefore, if we are passing a copy of the list, the searches will never reflect any changes (entries deleted or added).

PRoviding an instance to a provider, which can be the dream diary itself, we can always retrieve the entries as they are now.